Close #71
