+++ features/Elasticsearch.feature
@@ -6,7 +6,7 @@ Feature: Elasticsearch commands
     And I create the following document:
       | _id  | "chuon-chuon-kim" |
       | body | {}                |
-    When I run the command "es:get" with args:
+    When I run the command "es:indices:get" with args:
       | "&nyc-open-data.yellow-taxi" |
       | "chuon-chuon-kim"            |
     Then I should match stdout with "chuon-chuon-kim"
@@ -14,7 +14,19 @@ Feature: Elasticsearch commands
   @mappings
   Scenario: List ES indexes
     Given a collection "nyc-open-data":"green-taxi"
-    When I run the command "es:list-index" with flags:
+    When I run the command "es:indices:list" with flags:
       | --grep | "nyc-open-data" |
     Then I should match stdout with "yellow-taxi"
     And I should match stdout with "green-taxi"
+
+  @mappings
+  Scenario: Insert ES document
+    Given a collection "nyc-open-data":"green-taxi"
+    When I run the command "es:indices:insert" with:
+      | arg  | &nyc-open-data.yellow-taxi |         |
+      | flag | --id                       | kindred |
+      | flag | --body                     | {}      |
+    When I run the command "es:indices:get" with args:
+      | "&nyc-open-data.yellow-taxi" |
+      | "kindred"                    |
+    Then I should match stdout with "kindred"
